requirements.txt
livekit-agents
livekit-plugins-openai
livekit-plugins-deepgram
livekit-plugins-noise-cancellation
livekit-plugins-silero
livekit-plugins-cartesia
aiohttp
python-dotenv
